OS(framy): Re-enable TLP, fix amdgpu IO_PAGE_FAULT
I've been having the same issue for around a week now, and I thought
disabling TLP would fix it, but it didn't.
When disconnecting the dock, the screen would go gray, only showing the
mouse cursor, and switching to a TTY would work; when waking up from
sleep, the screen would go gray too, but this time I wouldn't be able to
switch to a TTY. In both cases the logs showed an IO_PAGE_FAULT error
from the amdgpu driver.

According to the Framework team, setting the option
`amdgpu.sg_display=0` and setting the VRAM allocation in the BIOS to
UMA_GAME_OPTIMIZED would fix the issue; I tried it, but that caused my
external display to not be detected, so I reverted the BIOS setting.

That actually fixed the issue!

This commit also re-enables TLP, as it was not the cause of the issue.